refactored for dynamic loading
