--- lib/active_record/transitions.rb
@@ -1,5 +1,0 @@
-require 'active_model/transitions'
-
-module ActiveRecord
-  Transitions = ActiveModel::Transitions
-end
