+++ lib/transitions/engine.rb
@@ -1,0 +1,4 @@
+module Transitions #:nodoc:
+  class Engine < ::Rails::Engine #:nodoc: 
+  end
+end
