+++ lib/transitions/exceptions.rb
@@ -1,0 +1,5 @@
+module Transitions
+  class InvalidTransition     < StandardError; end
+  class InvalidMethodOverride < StandardError; end
+end
+
