+++ lib/transitions/hooks.rb
@@ -1,0 +1,10 @@
+module Transitions
+  class Hooks
+    def self.init!
+      ActiveSupport.on_load(:active_record) do
+        require 'transitions/extension/transitions'
+        ::ActiveRecord::Base.send :include, Transitions::Extension
+      end
+    end
+  end
+end
