+++ lib/transitions/railtie.rb
@@ -1,0 +1,7 @@
+module Transitions
+  class Railtie < ::Rails::Railtie #:nodoc:
+    initializer 'transitions' do |app|
+      Transitions::Hooks.init!
+    end
+  end
+end
