+++ test/machine/test_machine.rb
@@ -1,8 +1,9 @@
 require "helper"
 
 class MachineTestSubject
-  include Transitions
+  include Transitions::Extension
 
+  include_state_machine
   state_machine :initial => :closed do
     state :open
     state :closed
