+++ test/state/test_state_predicate_method.rb
@@ -1,10 +1,11 @@
 require "helper"
 
 class Bus
-  include Transitions
+  include Transitions::Extension
   EXISTING_PREDICATE_METHOD_NAME = :driving?
   EXISTING_PREDICATE_METHOD_RETURN_VALUE = 'Am I driving?'
 
+  include_state_machine
   state_machine do
     state :parking
     state :driving
